+++ src/message_element_metadata_owner.cc
@@ -74,11 +74,11 @@ namespace grpc_labview {
 
     //---------------------------------------------------------------------
     //---------------------------------------------------------------------
-    void MessageElementMetadataOwner::RegisterMetadata(std::shared_ptr<EnumMetadata> requestMetadata)
+    void MessageElementMetadataOwner::RegisterMetadata(std::shared_ptr<EnumMetadata> metadata)
     {
         std::lock_guard<std::mutex> lock(_mutex);
 
-        _registeredEnumMetadata.insert({ requestMetadata->messageName, requestMetadata });
+        _registeredEnumMetadata.insert({ metadata->messageName, metadata });
     }
 
     //---------------------------------------------------------------------
@@ -146,41 +146,11 @@ namespace grpc_labview {
         metadata->clusterSize = AlignClusterOffset(clusterOffset, maxAlignmentRequirement);
     }
 
-    int MessageElementMetadataOwner::GetMaxAlignmentRequirement(LVMessageMetadataType elementType, bool elementIsRepeated)
-    {
-        int clusterOffset = AlignClusterOffset(clusterOffset, elementType, elementIsRepeated);
-        //int maxAlignmentRequirement = 0;
-        //element->clusterOffset = clusterOffset;
-        int elementSize = ClusterElementSize(elementType, elementIsRepeated);
-        clusterOffset += elementSize;       
-
-        return elementSize;
-    }
-
-    //---------------------------------------------------------------------
-    //---------------------------------------------------------------------
-    void MessageElementMetadataOwner::UpdateMetadataClusterLayout(std::shared_ptr<EnumMetadata>& metadata)
-    {
-        if (metadata->clusterSize != 0)
-        {
-            return;
-        }
-        int clusterOffset = 0;
-        int maxAlignmentRequirement = 0;
-        //maxAlignmentRequirement = GetMaxAlignmentRequirement((LVMessageMetadataType)9 /*TODO: Clean up*/, false /*element->isRepeated*/); // This should only be the enum definition, so the "repeated" part should get handled in the field of the message. Is this code required at all?
-        metadata->alignmentRequirement = maxAlignmentRequirement;
-        metadata->clusterSize = AlignClusterOffset(clusterOffset, maxAlignmentRequirement);
-    }
-
     //---------------------------------------------------------------------
     //---------------------------------------------------------------------
     void MessageElementMetadataOwner::FinalizeMetadata()
     {
         for (auto metadata: _registeredMessageMetadata)
-        {
-            UpdateMetadataClusterLayout(metadata.second);
-        }
-        for (auto metadata : _registeredEnumMetadata)
         {
             UpdateMetadataClusterLayout(metadata.second);
         }
